+++ src/OmniShadowMap.cpp
@@ -18,7 +18,7 @@ bool OmniShadowMap::init()
 
     // Initialize the 6 textures of the cubemap
     // These enums represent integers
-    for (size_t i {0}; i < 6; ++i)
+    for (int i {0}; i < 6; ++i)
         glTexImage2D(GL_TEXTURE_CUBE_MAP_POSITIVE_X + i, 0, GL_DEPTH_COMPONENT, m_shadowWidth, m_shadowHeight, 0, GL_DEPTH_COMPONENT, GL_FLOAT, nullptr);
 
     glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
